Implement statement 3
Albert: I don't know when Cheryl's birthday is, but I know that Bernard does not know too.

The statement3 takes as input a possible birthdate and returns true if Albert's statement is true for that birthdate.

Albert: After Cheryl told me the month of her birthdate, I didn't know her birthday. I don't know which day Cheryl told Bernard, but I know that for all of the possible dates, if Bernard is told that day, he wouldn't know the birthdate.